requirements.txt
Pillow
pyusb
